Update style-loader/mini-css-extract-plugin
Update style-loader and mini-css-extract-plugin
to the latest versions that are compatible
with webpack 4. The mini-css-extract-plugin
release notes recommend upgrading to style-loader
2.0.0 to avoid issues with esModule, so that
is why both are updated in the same commit.

The release notes for mini-css-extract-plugin 1.x
mention webpack 5 compatibility fixes, so this
update should simplify upgrading to webpack 5
when that time comes.